+++ resources/js/components/LocaleSelector.vue
@@ -5,7 +5,7 @@
       <i class="fa-solid fa-language"></i><span class="sr-only">{{ $t('app.select_locale') }}</span>
     </template>
     <b-dropdown-item
-      v-for="(label, locale) in locales"
+      v-for="(label, locale) in getSetting('enabled_locales')"
       :key="locale"
       @click="changeLocale(locale)"
       :active="locale === currentLocale">
@@ -15,22 +15,18 @@
 </template>
 
 <script>
-import { getLocaleList } from '../i18n';
 import env from './../env.js';
 import Base from '../api/base';
 import { mapActions, mapState } from 'pinia';
 import { useLocaleStore } from '../stores/locale';
 import { useLoadingStore } from '../stores/loading';
+import { useSettingsStore } from '../stores/settings';
 
 export default {
 
   computed: {
-
-    locales () {
-      return getLocaleList();
-    },
-
-    ...mapState(useLocaleStore, ['currentLocale'])
+    ...mapState(useLocaleStore, ['currentLocale']),
+    ...mapState(useSettingsStore, ['getSetting'])
   },
   methods: {
     ...mapActions(useLocaleStore, ['setLocale']),