+++ resources/js/stores/auth.js
@@ -1,6 +1,5 @@
 import { defineStore } from 'pinia';
 import auth from '../api/auth';
-import { setTimeZone } from '../i18n';
 import PermissionService from '../services/PermissionService';
 import _ from 'lodash';
 import { useLocaleStore } from './locale';
@@ -18,11 +17,6 @@ export const useAuthStore = defineStore('auth', {
     login: async function (credentials, method) {
       await auth.login(credentials, method);
       await this.getCurrentUser();
-
-      if (this.currentUser.user_locale !== null) {
-        const locale = useLocaleStore();
-        await locale.setCurrentLocale(this.currentUser.user_locale);
-      }
     },
 
     async getCurrentUser () {
@@ -32,7 +26,11 @@ export const useAuthStore = defineStore('auth', {
       }
       // set timezone of i18n, if user not logged in use undefined to set timezone to local system timezone
       const locale = useLocaleStore();
-      setTimeZone(locale.i18n, currentUser == null ? undefined : currentUser.timezone);
+      locale.setTimezone(currentUser == null ? undefined : currentUser.timezone);
+
+      if (currentUser != null && currentUser.user_locale != null) {
+        await locale.setCurrentLocale(currentUser.user_locale);
+      }
 
       this.setCurrentUser(currentUser);
     },
@@ -43,9 +41,8 @@ export const useAuthStore = defineStore('auth', {
       // logout successfull, clear current user
       this.setCurrentUser(null, false);
       // reset timezone of i18n to use local system timezone
-
       const locale = useLocaleStore();
-      setTimeZone(locale.i18n, undefined);
+      locale.setTimezone(undefined);
 
       return response;
     },