+++ tests/Fixtures/Locales/custom/de/app.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'key_1' => 'neuer_wert_1',
+];