+++ tests/Frontend/components/Room/RoomFile.spec.js
@@ -360,7 +360,7 @@ describe('RoomFile', () => {
 
     const request = mockAxios.request('/api/v1/rooms/123-456-789/files');
 
-    view.vm.uploadFile({ target: { files: [file] } });
+    view.vm.uploadFile(file);
 
     await request.wait();
     expect(request.config.headers['Content-Type']).toBe('multipart/form-data');